WeaponAttacker -> ShortRangeWeapon
