+++ Assets/SSP/Prefabs/Player.prefab
@@ -1400,7 +1400,7 @@ GameObject:
   - component: {fileID: 33137232365910524}
   - component: {fileID: 136212414492031444}
   - component: {fileID: 23471199039844016}
-  - component: {fileID: 114097779763349048}
+  - component: {fileID: 114205188047944820}
   m_Layer: 0
   m_Name: bat
   m_TagString: Untagged
@@ -3028,7 +3028,7 @@ Transform:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1340052829251104}
-  m_LocalRotation: {x: -0.093746535, y: 0.6978551, z: 0.707469, w: 0.060806695}
+  m_LocalRotation: {x: -0.09374654, y: 0.6978552, z: 0.70746905, w: 0.0608067}
   m_LocalPosition: {x: -0.05543403, y: 0.0017306843, z: 0.098653734}
   m_LocalScale: {x: 1.0000004, y: 1.0000006, z: 1.0000005}
   m_Children:
@@ -5287,20 +5287,6 @@ MonoBehaviour:
   offset: {x: 0, y: 2, z: -3}
   cameraRotationSpeed: 100
   pim: {fileID: 114858071444870302}
---- !u!114 &114097779763349048
-MonoBehaviour:
-  m_ObjectHideFlags: 1
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1519376701410896}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b82852eb7b1af244b8227b344a84b049, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  damageAmount: 10
-  hitDetectionTimeOffset: 0.2
-  hitDetectionDuration: 0.6
 --- !u!114 &114150559372493268
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -5331,6 +5317,20 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   avoidStartTime: 0.1
   avoidDuration: 0.5
+--- !u!114 &114205188047944820
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1519376701410896}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7fb53c3547122084c8b1055181f33790, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  damageAmount: 10
+  hitDetectionTimeOffset: 0.2
+  hitDetectionDuration: 0.6
 --- !u!114 &114247152544189534
 MonoBehaviour:
   m_ObjectHideFlags: 1
