--- Assets/SSP/Scripts/Weapons/WeaponAttacker.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: b82852eb7b1af244b8227b344a84b049
-timeCreated: 1503677377
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
